+++ src/main/java/com/EIPplatform/model/dto/report/report05/airemmissionmanagement/airautoqcvnexceedance/AirAutoQcvnExceedanceDTO.java
@@ -3,7 +3,7 @@ package com.EIPplatform.model.dto.report.report05.airemmissionmanagement.airauto
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
+import java.lang.Double;
 
 @Data
 @Builder
@@ -20,5 +20,5 @@ public class AirAutoQcvnExceedanceDTO {
 
     Integer qcvnLimitValue;
 
-    BigDecimal exceedRatioPercent;
+    Double exceedRatioPercent;
 }
