+++ src/main/java/com/EIPplatform/model/dto/report/report05/wastemanagement/exportedhwstat/ExportedHwStatUpdateDTO.java
@@ -3,7 +3,7 @@ package com.EIPplatform.model.dto.report.report05.wastemanagement.exportedhwstat
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
+import java.lang.Double;
 
 @Data
 @Builder
@@ -17,7 +17,7 @@ public class ExportedHwStatUpdateDTO {
 
     String baselCode;
 
-    BigDecimal volumeKg;
+    Double volumeKg;
 
     String transporterOrg;
