+++ src/main/java/com/EIPplatform/model/dto/report/report05/wastemanagement/industrialsolidwastestat/IndustrialSolidWasteStatDTO.java
@@ -3,7 +3,7 @@ package com.EIPplatform.model.dto.report.report05.wastemanagement.industrialsoli
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
+import java.lang.Double;
 
 @Data
 @NoArgsConstructor
@@ -13,7 +13,7 @@ import java.math.BigDecimal;
 public class IndustrialSolidWasteStatDTO {
     Long industrialId;
     String wasteGroup;
-    BigDecimal volumeCy;
+    Double volumeCy;
     String receiverOrg;
-    BigDecimal volumePy;
+    Double volumePy;
 }
